refactor: adjust corresponding packages
