+++ src/allocation/adapters/__init__.py
@@ -1,0 +1,5 @@
+"""
+Adapters is a nod to the ports and adapters terminology. This will fill up with any other abstractions around
+external I/O (e.g., a redis_client.py). Strictly speaking, you would call these secondary adapters or driven
+adapters, or sometimes inward-facing adapters.
+"""
