+++ tests/mocks.py
@@ -1,9 +1,9 @@
 """
 This module describes all shared mocks.
 """
+from allocation.adapters.repository import SqlAlchemyRepository
 from allocation.dependencies import get_batch_repository, get_session
 from allocation.main import app
-from allocation.repository.repository import SqlAlchemyRepository
 
 
 def override_session(session):
